day_2/project_gallery/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
drf-yasg==1.21.3